Initial cockpit page
We can't use the dev server, as pages need to be served through
cockpit-ws. So replace with watch mode.

Add a starter-kit like demo of the cockpit API by showing the hostname.

Caveat: `import`ing cockpit (as we do in current webpack projecs) does
*not* work. This needs some more thought.

Needs `unsafe-inline` for PatternFly's CSS.

This is currently blocked on https://github.com/cockpit-project/cockpit/issues/15288
as cockpit-bridge does not properly load paths that contain @.